Rename `cupyx.scatter` submodule
