+++ cupyx/__init__.py
@@ -1,9 +1,9 @@
 # "NOQA" to suppress flake8 warning
 from cupyx.rsqrt import rsqrt  # NOQA
 from cupyx.runtime import get_runtime_info  # NOQA
-from cupyx.scatter import scatter_add  # NOQA
-from cupyx.scatter import scatter_max  # NOQA
-from cupyx.scatter import scatter_min  # NOQA
+from cupyx._scatter import scatter_add  # NOQA
+from cupyx._scatter import scatter_max  # NOQA
+from cupyx._scatter import scatter_min  # NOQA
 
 from cupyx import linalg  # NOQA
 from cupyx import time  # NOQA
